ref: refactor policy adapter

+++ internal/usecase/config/delete.go
@@ -4,8 +4,13 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/sergeyslonimsky/elara/internal/auth"
 	"github.com/sergeyslonimsky/elara/internal/domain"
 )
+
+type deleteEnforcer interface {
+	Enforce(subject, domain, object, action string) (bool, error)
+}
 
 type configDeleter interface {
 	Delete(ctx context.Context, path, namespace string) (int64, error)
@@ -16,17 +21,32 @@ type deleteWatchNotifier interface {
 }
 
 type DeleteUseCase struct {
-	configs configDeleter
-	watch   deleteWatchNotifier
+	enforcer deleteEnforcer
+	configs  configDeleter
+	watch    deleteWatchNotifier
 }
 
-func NewDeleteUseCase(configs configDeleter, watch deleteWatchNotifier) *DeleteUseCase {
-	return &DeleteUseCase{configs: configs, watch: watch}
+func NewDeleteUseCase(enforcer deleteEnforcer, configs configDeleter, watch deleteWatchNotifier) *DeleteUseCase {
+	return &DeleteUseCase{enforcer: enforcer, configs: configs, watch: watch}
 }
 
 func (uc *DeleteUseCase) Execute(ctx context.Context, path, namespace string) error {
 	if namespace == "" {
 		return domain.NewValidationError("namespace", "namespace is required")
+	}
+
+	claims, ok := auth.ClaimsFromContext(ctx)
+	if !ok {
+		return domain.ErrUnauthorized
+	}
+
+	allowed, err := uc.enforcer.Enforce(claims.Email, namespace, "config", "write")
+	if err != nil {
+		return fmt.Errorf("enforce: %w", err)
+	}
+
+	if !allowed {
+		return domain.ErrForbidden
 	}
 
 	rev, err := uc.configs.Delete(ctx, path, namespace)
